Add code to output particle potentials. Enable with -DOUTPUT_PARTICLE_POTENTIAL.
When enabled, an extra particle array is generated and, at each move_fine step, updated with the potential
from the parent cell. At output steps, the array is written out after the mass.

In writing this code, I have also unified some of the logic which determines how large arrays are when
scattering/gathering particle properties between MPI nodes. This modification applies whether or not
OUTPUT_PARTICLE_POTENTIAL is defined, but should be transparent.

To help readers make sense of whether OUTPUT_PARTICLE_POTENTIAL is enabled or not, I also added
a line to the header.txt file which states transparently what particle fields are present.


Former-commit-id: 858c93b5a16378d9e81777bf9df4e3912ee76563